separate email logic from user logic, created new src/email.rs file
